+++ src/services/index.ts
@@ -3,6 +3,7 @@ import {
   startTunnelRegistryServer,
 } from '../lib/tunnel/tunnel-registry-server.js';
 import * as afc from './ios/afc/index.js';
+import * as crashReports from './ios/crash-reports/index.js';
 import * as diagnostics from './ios/diagnostic-service/index.js';
 import * as houseArrest from './ios/house-arrest/index.js';
 import * as mobileImageMounter from './ios/mobile-image-mounter/index.js';
@@ -12,6 +13,7 @@ import * as tunnel from './ios/tunnel-service/index.js';
 import * as webinspector from './ios/webinspector/index.js';
 
 export {
+  crashReports,
   diagnostics,
   houseArrest,
   mobileImageMounter,